requirements.txt
Flask
SQLAlchemy-Utils
Flask-SQLAlchemy
flask_debugtoolbar
colorama
termcolor
furl
binaryornot
pytz